renaming and docs
